refactor(widget): relocate service implementation with modular helper classes
The widget service has been restructured by moving the main WidgetService
class to a dedicated directory and splitting its functionality across
multiple focused helper classes. This improves separation of concerns
and makes the codebase more maintainable through better organization.

The following helper classes were created:
- WidgetDataAggregator for handling data aggregation logic
- WidgetToggleHelper for managing task/habit toggling operations
- WidgetFeedbackManager for handling user feedback displays
- widgetBackgroundCallback extracted to isolated handler

All imports have been updated to reference the new file locations under
the services/widget_service/widget_service.dart path structure.

BREAKING CHANGE: WidgetService location changed from
src/lib/core/application/features/widget/services/widget_service.dart to
src/lib/core/application/features/widget/services/widget_service/widget_service.dart
requiring import path updates in dependent modules.
